WEB/[Simple KB3]/src/requirements.txt
Flask==3.1.0
Requests==2.32.3
PyJWT==2.3.0
PyYAML
pathlib
cryptography